pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keycase-agent-sdk"
version = "0.1.0b1"
description = "Python SDK for Keycase Agent execution and keyword integration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Keycase Team", email = "support@keycase.io"}
]
maintainers = [
    {name = "Keycase Team", email = "support@keycase.io"}
]
keywords = [
    "automation",
    "testing",
    "keyword-driven",
    "test-automation",
    "sdk",
    "agent",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "websocket-client>=1.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "types-requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/TharassKeycase/keycase-python-sdk"
Documentation = "https://docs.keycase.io/python-agent"
Repository = "https://github.com/TharassKeycase/keycase-python-sdk"
Issues = "https://github.com/TharassKeycase/keycase-python-sdk/issues"
Changelog = "https://github.com/TharassKeycase/keycase-python-sdk/blob/main/CHANGELOG.md"

[project.scripts]
keycase-agent = "keycase_agent:start_agent"

[tool.setuptools.packages.find]
where = ["."]
include = ["keycase_agent*"]

[tool.setuptools.package-data]
keycase_agent = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["keycase_agent"]
skip = [".venv", "venv", "build", "dist"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["keycase_agent"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
